feat: new Bun-only workspace + Tailwind v4 UI; move legacy to old-plugin; add figma-kit
