+++ manifest.json
@@ -1,9 +1,1 @@
-{
-    "api": "1.0.0",
-    "editorType": ["figma"],
-    "id": "1351760630951154540",
-    "name": "Theme Engine",
-    "main": "build/main.js",
-    "ui": "build/ui.js",
-    "networkAccess": { "allowedDomains": ["none"] }
-}
+{"api":"1.0.0","editorType":["figma"],"id":"1351760630951154540","name":"Theme Engine","main":"build/main.js","ui":"build/ui.js"}
