+++ old-plugin/scripts/update-manifest-ui.mjs
