+++ old-plugin/src/components/gradient-preview.tsx
