+++ old-plugin/src/components/primitives-tab/color-name-input.tsx
