+++ old-plugin/src/components/primitives-tab/color-picker.tsx
