+++ old-plugin/src/defaults.ts
