+++ old-plugin/src/hooks/defaults/aliases.ts
