+++ old-plugin/src/hooks/defaults/theme-color.ts
