+++ old-plugin/src/hooks/defaults/theme.ts
