+++ old-plugin/src/styles.css.d.ts
@@ -1,0 +1,23 @@
+declare const styles: {
+  readonly "_tabs_61qsz_1": string;
+  readonly "_opacityInput_69uub_84": string;
+  readonly "_divider_69uub_109": string;
+  readonly "_textboxColor_69uub_1": string;
+  readonly "_hexColorInput_69uub_80": string;
+  readonly "_border_s66c9_8": string;
+  readonly "hue-slider-bar": string;
+  readonly "tab-content": string;
+  readonly "build-button": string;
+  readonly "gradientMove": string;
+  readonly "import-button": string;
+  readonly "fadeInOut": string;
+  readonly "input": string;
+  readonly "selected-theme-color": string;
+  readonly "swatch-button": string;
+  readonly "bg-checkerboard": string;
+  readonly "scrollbar-hide": string;
+  readonly "toggled-theme-color": string;
+  readonly "icon-toggle-styles": string;
+};
+export = styles;
+
