+++ old-plugin/src/variable-collection.ts
