+++ old-plugin/ui/index.html
@@ -3,12 +3,11 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Theme Engine (New UI)</title>
-    <link rel="stylesheet" href="/src/app.css" />
+    <title>Theme Engine UI</title>
   </head>
-  <body class="bg-white text-black">
+  <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="../src/react-ui/main.tsx"></script>
   </body>
   </html>
