+++ package.json
@@ -1,76 +1,68 @@
 {
-  "dependencies": {
-    "@create-figma-plugin/ui": "3.1.0",
-    "@create-figma-plugin/utilities": "3.1.0",
-    "@material/material-color-utilities": "^0.2.7",
-    "@preact/signals": "^1.2.3",
-    "@radix-ui/react-icons": "^1.3.0",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.0.0",
-    "complex.js": "^2.1.1",
-    "figx": "^0.1.3",
-    "javascript-natural-sort": "^0.7.1",
-    "lodash": "^4.17.21",
-    "lucide-react": "^0.293.0",
-    "mathjs": "^12.4.1",
-    "nanoid": "^5.0.4",
-    "nanoid-dictionary": "^5.0.0-beta.1",
-    "npm": "^9.8.1",
-    "preact": "^10.19.2",
-    "prettier": "^3.2.2",
-    "prettier-plugin-tailwindcss": "^0.5.11",
-    "react-dom": "^19.0.0",
-    "typed-function": "^4.1.1",
-    "uuid": "^9.0.0",
-    "validator": "^13.11.0",
-    "zod": "^3.22.4",
-    "zod-validation-error": "^2.1.0",
-    "zustand": "^4.4.4"
-  },
-  "devDependencies": {
-    "@create-figma-plugin/build": "^2.5.1",
-    "@create-figma-plugin/tsconfig": "^2.5.1",
-    "@figma/plugin-typings": "^1.88.0",
-    "@types/bun": "^1.0.5",
-    "@types/lodash": "^4.14.196",
-    "@types/react": "^19.0.0",
-    "@types/uuid": "^9.0.2",
-    "@types/validator": "^13.11.9",
-    "@typescript-eslint/eslint-plugin": "^7.1.0",
-    "@typescript-eslint/parser": "^7.1.0",
-    "autoprefixer": "^10.4.14",
-    "css-loader": "^6.8.1",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "postcss": "^8.4.25",
-    "postcss-cli": "^10.1.0",
-    "@vitejs/plugin-react": "^4.3.0",
-    
-    "react": "^19.0.0",
-    "style-loader": "^3.3.3",
-    "tailwind-merge": "^1.14.0",
-    "tailwindcss": "^3.4.1",
-    "tailwindcss-animate": "^1.0.7",
-    "typescript": "^5.3.3",
-    "vite": "^5.4.0"
-  },
-  "scripts": {
-    "build": "npm run build:plugin && npm run build:ui",
-    "build:plugin": "build-figma-plugin --typecheck --minify",
-    "build:ui": "vite build",
-    "dev": "vite",
-    "ui:activate": "node scripts/update-manifest-ui.mjs",
-    "watch": "build-figma-plugin --typecheck --watch",
-    "tailwind": "postcss ./src/styles.css -o ./src/dist/tailwind.css --watch"
-  },
-  "figma-plugin": {
-    "editorType": [
-      "figma"
+    "packageManager": "bun@1",
+    "type": "module",
+    "engines": {
+        "bun": ">=1.1.0"
+    },
+    "engineStrict": true,
+    "workspaces": [
+        "packages/*"
     ],
-    "id": "1351760630951154540",
-    "name": "Theme Engine",
-    "main": "src/main.ts",
-    "ui": "src/ui.tsx"
-  }
+    "dependencies": {
+        "@create-figma-plugin/ui": "3.1.0",
+        "@create-figma-plugin/utilities": "3.1.0",
+        "@material/material-color-utilities": "^0.2.7",
+        "@preact/signals": "^1.2.3",
+        "class-variance-authority": "^0.7.0",
+        "clsx": "^2.0.0",
+        "complex.js": "^2.1.1",
+        "figx": "^0.1.3",
+        "figma-kit": "workspace:*",
+        "javascript-natural-sort": "^0.7.1",
+        "lodash": "^4.17.21",
+        "mathjs": "^12.4.1",
+        "nanoid": "^5.0.4",
+        "nanoid-dictionary": "^5.0.0-beta.1",
+        "preact": "^10.19.2",
+        "prettier": "^3.2.2",
+        "prettier-plugin-tailwindcss": "^0.5.11",
+        "react-dom": "^19.0.0",
+        "typed-function": "^4.1.1",
+        "uuid": "^9.0.0",
+        "validator": "^13.11.0",
+        "zod": "^3.22.4",
+        "zod-validation-error": "^2.1.0",
+        "zustand": "^4.4.4"
+    },
+    "devDependencies": {
+        "@create-figma-plugin/build": "^2.5.1",
+        "@create-figma-plugin/tsconfig": "^2.5.1",
+        "@figma/plugin-typings": "^1.88.0",
+        "@types/bun": "^1.0.5",
+        "@types/lodash": "^4.14.196",
+        "@types/react": "^19.0.0",
+        "@types/react-dom": "^19.0.0",
+        "@types/uuid": "^9.0.2",
+        "@types/validator": "^13.11.9",
+        "@typescript-eslint/eslint-plugin": "^7.1.0",
+        "@typescript-eslint/parser": "^7.1.0",
+        "eslint": "^8.57.0",
+        "eslint-plugin-react": "^7.33.2",
+        "eslint-plugin-react-hooks": "^4.6.0",
+        "@vitejs/plugin-react": "^4.3.0",
+        "@tailwindcss/vite": "^4.1.11",
+        "react": "^19.0.0",
+        "react-refresh": "^0.14.0",
+        "tailwind-merge": "^1.14.0",
+        "tailwindcss": "^4.1.11",
+        "tailwindcss-animate": "^1.0.7",
+        "typescript": "^5.3.3",
+        "vite": "^5.4.0"
+    },
+    "scripts": {
+        "build": "bun run build:worker && bun run build:ui",
+        "build:ui": "bunx vite build",
+        "build:worker": "bun build src/main.ts --outfile=build/main.js --format=cjs",
+        "dev": "bunx vite"
+    }
 }
