+++ packages/figma-kit/src/index.ts
@@ -1,0 +1,3 @@
+export { Button } from './components/Button';
+
+
