+++ packages/figma-kit/tsconfig.json
@@ -1,0 +1,15 @@
+{
+    "compilerOptions": {
+        "target": "ES2020",
+        "module": "ESNext",
+        "jsx": "react-jsx",
+        "declaration": true,
+        "outDir": "dist",
+        "moduleResolution": "Bundler",
+        "skipLibCheck": true,
+        "strict": true
+    },
+    "include": ["src"]
+}
+
+
