+++ ui/src/App.tsx
@@ -1,0 +1,16 @@
+/** @jsxImportSource react */
+import React from 'react';
+
+export default function App() {
+    return (
+        <div className="p-4 space-y-4">
+            <h1 className="text-xl font-semibold">Theme Engine (New UI)</h1>
+            <p className="text-sm text-gray-600">Tailwind v4 + React 19 scaffold</p>
+            <div className="flex gap-2">
+                <button className="rounded bg-black text-white px-3 py-1">Button</button>
+            </div>
+        </div>
+    );
+}
+
+
