+++ ui/src/app.css
@@ -1,0 +1,40 @@
+@import 'tailwindcss';
+
+:root {
+    --radius: 0.5rem;
+
+    /* Core theme tokens (expand as needed) */
+    --background: 0 0% 100%;
+    --foreground: 222.2 84% 4.9%;
+
+    --primary: 221.2 83.2% 53.3%;
+    --primary-foreground: 210 40% 98%;
+
+    --secondary: 210 40% 96.1%;
+    --secondary-foreground: 222.2 47.4% 11.2%;
+
+    --muted: 210 40% 96.1%;
+    --muted-foreground: 215.4 16.3% 46.9%;
+
+    --accent: 210 40% 96.1%;
+    --accent-foreground: 222.2 47.4% 11.2%;
+
+    --destructive: 0 84.2% 60.2%;
+    --destructive-foreground: 210 40% 98%;
+
+    --border: 214.3 31.8% 91.4%;
+    --input: 214.3 31.8% 91.4%;
+    --ring: 221.2 83.2% 53.3%;
+
+    --fig-blue: 24 86% 54%;
+    --gridlines: 0 0% 50% / 0.25;
+
+    --spacing-172: 172px;
+    --spacing-344: 344px;
+}
+
+@utility container-padded {
+    padding-inline: 2rem;
+}
+
+
