+++ ui/src/main.tsx
@@ -1,14 +1,14 @@
+/** @jsxImportSource react */
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App';
-import '../styles.css';
 
 const container = document.getElementById('root');
-if (!container) {
-	throw new Error('Root container not found');
-}
-
-const root = createRoot(container);
-root.render(<App />);
+const root = createRoot(container!);
+root.render(
+    <React.StrictMode>
+        <App />
+    </React.StrictMode>,
+);
