+++ vite.config.ts
@@ -1,10 +1,12 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import tailwindcss from '@tailwindcss/vite';
 
 export default defineConfig({
 	root: 'ui',
 	plugins: [
-		react(),
+        tailwindcss(),
+        react(),
 	],
 	build: {
 		outDir: 'build/ui',
